Use numpy 128 bits float (#47)
(cherry picked from commit fe913ee6154bc55d805992ab9a834f6c1f7c27dc)

+++ dnn/setting.py
@@ -7,20 +7,20 @@ import numpy as np
 # 重みの作成
 #一様分布
 def unif(i_node, o_node):
-    weight = np.random.rand(i_node * o_node) * 2 - 1
-    weight = weight.reshape(o_node, i_node)
+    weight = np.random.rand(i_node * o_node).astype('f16') * 2 - 1
+    weight = weight.reshape(o_node, i_node).astype('f16')
     return weight
 
 #正規分布(xivier)
 def xivier(i_node, o_node):
-    weight = np.random.normal(loc=0.0, scale=1 / math.sqrt(i_node), size=i_node * o_node)
-    weight = weight.reshape(o_node, i_node)
+    weight = np.random.normal(loc=0.0, scale=1 / math.sqrt(i_node), size=i_node * o_node).astype('f16')
+    weight = weight.reshape(o_node, i_node).astype('f16')
     return weight
 
 #正規分布(he)
 def he(i_node, o_node):
-    weight = np.random.normal(loc=0.0, scale=math.sqrt(2 / i_node), size=i_node * o_node)
-    weight = weight.reshape(o_node, i_node)
+    weight = np.random.normal(loc=0.0, scale=math.sqrt(2 / i_node), size=i_node * o_node).astype('f16')
+    weight = weight.reshape(o_node, i_node).astype('f16')
     return weight
 
 # y
@@ -28,7 +28,7 @@ def ynet(layer):
     length = len(layer)
     net = list("-")
     for i in range(1, length):
-        y = np.ones(layer[i])
+        y = np.ones(layer[i], dtype=np.float128)
         net.append(y)
     return net
 
@@ -36,7 +36,7 @@ def ynet(layer):
 def znet(layer):
     net = list()
     for i in range(len(layer)):
-        x = np.zeros(layer[i])
+        x = np.zeros(layer[i], dtype=np.float128)
         net.append(x)
     return net
